requirements.txt
asgiref==3.7.2
black==23.7.0
boto3==1.28.85
botocore==1.31.85
click==8.1.6
colorama==0.4.6
Django==4.2.4
django-autoslug==1.9.9
django-cities==0.6.1
django-cities-light==3.9.2
django-ckeditor==6.7.0
django-countries==7.5.1
django-js-asset==2.1.0
django-model-utils==4.3.1
django-storages==1.14.2
jmespath==1.0.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.1.0
# Pillow=8.0.0
platformdirs==3.10.0
progressbar2==4.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-utils==3.7.0
pytz==2023.3
pyuca==1.2
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
Unidecode==1.3.6
urllib3==2.0.7
whitenoise==6.5.0
